+++ paddle_billing/Entities/Subscriptions/SubscriptionItemsWithPrice.py
@@ -1,5 +1,5 @@
 from __future__ import annotations
-from dataclasses import asdict, dataclass
+from dataclasses import dataclass
 
 from paddle_billing.Entities.Subscriptions.SubscriptionNonCatalogPrice import SubscriptionNonCatalogPrice
 from paddle_billing.Entities.Subscriptions.SubscriptionNonCatalogPriceWithProduct import (
@@ -18,6 +18,3 @@ class SubscriptionItemsWithPrice:
             price=data["price"],
             quantity=data["quantity"],
         )
-
-    def get_parameters(self) -> dict:
-        return asdict(self)