+++ paddle_billing/Resources/NotificationSettings/Operations/CreateNotificationSetting.py
@@ -1,18 +1,16 @@
-from dataclasses import asdict, dataclass
+from dataclasses import dataclass
 
+from paddle_billing.Operation import Operation
 from paddle_billing.Undefined import Undefined
 from paddle_billing.Entities.Events import EventTypeName
 from paddle_billing.Entities.NotificationSettings import NotificationSettingType
 
 
 @dataclass
-class CreateNotificationSetting:
+class CreateNotificationSetting(Operation):
     description: str
     destination: str
     subscribed_events: list[EventTypeName]
     type: NotificationSettingType
     include_sensitive_fields: bool
     api_version: int | Undefined = Undefined()
-
-    def get_parameters(self) -> dict:
-        return asdict(self)