+++ paddle_billing/Resources/Reports/Operations/CreateReport.py
@@ -1,13 +1,15 @@
 from dataclasses import dataclass, field
 from abc import ABC, abstractmethod
 
+from paddle_billing.Operation import Operation
+
 from paddle_billing.Entities.Reports import ReportType, ReportFilter
 
 from paddle_billing.Exceptions.SdkExceptions.InvalidArgumentException import InvalidArgumentException
 
 
 @dataclass
-class CreateReport(ABC):
+class CreateReport(Operation, ABC):
     type: ReportType
     filters: list[ReportFilter] = field(default_factory=list)
 
@@ -21,11 +23,11 @@ class CreateReport(ABC):
 
             raise InvalidArgumentException.array_contains_invalid_types("filters", allowed_type_names, invalid_items)
 
-    def get_parameters(self) -> dict:
+    def to_json(self) -> dict:
         parameters = {"type": self.type}
 
         if self.filters is not None and self.filters != []:
-            parameters.update({"filters": [filter_.get_parameters() for filter_ in self.filters]})
+            parameters.update({"filters": [filter_ for filter_ in self.filters]})
 
         return parameters