+++ paddle_billing/Resources/Transactions/Operations/PreviewTransactionByAddress.py
@@ -1,5 +1,6 @@
-from dataclasses import asdict, dataclass
+from dataclasses import dataclass
 
+from paddle_billing.Operation import Operation
 from paddle_billing.Undefined import Undefined
 from paddle_billing.Entities.Shared import AddressPreview, CurrencyCode
 from paddle_billing.Entities.Transactions import (
@@ -9,13 +10,10 @@ from paddle_billing.Entities.Transactions import (
 
 
 @dataclass
-class PreviewTransactionByAddress:
+class PreviewTransactionByAddress(Operation):
     address: AddressPreview
     items: list[TransactionItemPreviewWithPriceId | TransactionItemPreviewWithNonCatalogPrice]
     customer_id: str | None | Undefined = Undefined()
     currency_code: CurrencyCode | Undefined = Undefined()
     discount_id: str | None | Undefined = Undefined()
     ignore_trials: bool | Undefined = Undefined()
-
-    def get_parameters(self) -> dict:
-        return asdict(self)